pyproject.toml
[tool.poetry]
name = "lightroot"
version = "0.1.0"
description = ""
authors = ["Vikash Pathak <vikash@sellcx.com>"]
readme = "README.md"
packages = [
    { include = "core" },
    { include = "adapters" },
    { include = "nomad" }
]

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.1.0"
openai = "^1.77.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
